Ensure activesupport works with Ruby versions older than 2.2.2
activesupport 5 drops support for older Ruby versions, but we still want
to allow them. Unfortunately, RubyGems doesn't seem to take required
Ruby versions into account when selecting versions of Gems to install,
and will attempt to install AS 5 and break in Ruby < 2.2.2.

Work around the issue by manually locking the AS version in the Gemfile
if necessary. This will only help when setting up the development
environment, but not when depending on the Gem. Unfortunately there is
little we can do about it other than document that requirement, so that
the most external project can always include the version lock in their
Gemfile.